+++ package.json
@@ -23,6 +23,7 @@
     "react-test-renderer": "^16.12.0",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
+    "redux-saga": "^1.1.3",
     "reselect": "^4.0.0"
   },
   "scripts": {
