+++ src/components/Auth/index.js
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from "prop-types";
 import { makeStyles } from '@material-ui/core/styles';
 import Grid from '@material-ui/core/Grid';
-import {LoginForm} from './LoginForm';
+import { LoginForm } from '../Login/LoginForm';
+import { SignupForm } from '../Signup/SignupForm';
 import { Logo } from 'loft-taxi-mui-theme';
 
 const BackgroundURL = '../background.jpg';
@@ -20,7 +21,7 @@ const useStyles = makeStyles(theme => ({
   }
 }));
 
-export const Login = () => {
+export const Auth = ({isAuth}) => {
   const classes = useStyles();
 
   return (
@@ -34,13 +35,13 @@ export const Login = () => {
             <Logo white animated />
           </Grid>
           <Grid item xs={3}>
-            <LoginForm/>
+            {isAuth ? <LoginForm/> : <SignupForm/>}
           </Grid>
        </Grid>
     </div>
   );
 }
 
-Login.propTypes = {
-  setRoute: PropTypes.func
+Auth.propTypes = {
+  isAuth: PropTypes.bool
 };
