+++ src/components/General/AppRouter.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import { Map } from '../Map';
-import { Login } from '../Login';
-import { Signup } from '../Signup';
+import { Auth } from '../Auth';
 import { Profile } from '../Profile';
-import { getAuth } from '../../modules/main'
-import {shallowEqual, useSelector } from 'react-redux'
+import { getAuth } from '../../modules/auth';
+import {shallowEqual, useSelector } from 'react-redux';
 
 export const AppRouter = () => {
   const loginPath = '/login';
@@ -22,8 +21,8 @@ export const AppRouter = () => {
 
   return (
     <Switch>
-      <Route path={'/login'} component={Login}/>
-      <Route path={'/signup'} component={Signup}/>
+      <Route path={'/login'} component={() => <Auth isAuth={true} />}/>
+      <Route path={'/signup'} component={() => <Auth isAuth={false} />}/>
       <PrivateRoute path={'/map'} component={Map}/>
       <PrivateRoute path={'/profile'} component={Profile}/>
       <Redirect from={'/'} to={'login'}/>
