+++ src/modules/index.js
@@ -1,7 +1,17 @@
 
 import { combineReducers } from 'redux';
-import main from './main';
+import address from './address';
+import auth from './auth';
+import card from './card';
+import register from './register';
+import route from './route';
+// import main from './main';
 
 export default combineReducers({
-  main,
+  // main
+  address,
+  auth,
+  card,
+  register,
+  route
 });
